Suppress console for Windows build
Do not show a new console when the executable is started outside of a console. The stdout and stderr are still visible when the user starts scroom.exe from the command line.

This is sufficient for most users, and for extra debugging we can ask the users encountering a problem to open scroom.exe from the command line.